# core-service/src/main/resources/openapi/api-core-v1.yml
openapi: 3.0.3
info:
  title: RouteCraft Core API
  description: Полная спецификация Core Service (Маршруты, Пользователи, Избранное, Файлы)
  version: 1.0.0

servers:
  - url: /api
    description: Core Service

tags:
  - name: auth
    description: Аутентификация и управление сессиями
  - name: users
    description: Профиль пользователя
  - name: routes
    description: Основной CRUD маршрутов
  - name: favorites
    description: Сохраненные маршруты
  - name: reviews
    description: Отзывы и рейтинги
  - name: admin
    description: Панель администратора

paths:
  /auth/register:
    post:
      tags:
        - auth
      summary: Регистрация нового пользователя
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'

  /auth/login:
    post:
      tags:
        - auth
      summary: Вход в систему
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /auth/refresh:
    post:
      tags:
        - auth
      summary: Обновление токенов
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /auth/logout:
    post:
      tags:
        - auth
      summary: Выход из системы
      operationId: logout
      responses:
        '204':
          description: Успешный выход
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /auth/check-email-exists:
    get:
      tags:
        - auth
      summary: Проверить, зарегистрирован ли email
      operationId: checkEmailExists
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Результат проверки email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckEmailExistsResponse'
        '400':
          $ref: '#/components/responses/ValidationError'

  /auth/validate:
    get:
      tags:
        - auth
      summary: Validate access token for API Gateway
      operationId: validateToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthValidationResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /users/me:
    get:
      tags:
        - users
      summary: Получить данные текущего профиля
      operationId: getCurrentUser
      responses:
        '200':
          description: Данные пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /routes:
    get:
      tags:
        - routes
      summary: Получить список маршрутов с поиском и пагинацией
      operationId: getRoutes
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 1)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: search
          in: query
          description: Поисковая строка по названию или описанию маршрута
          schema:
            type: string
      responses:
        '200':
          description: Список маршрутов с метаданными пагинации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoutesResponse'

    post:
      tags:
        - routes
      summary: Создать новый маршрут
      operationId: createRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRouteRequest'
      responses:
        '201':
          description: ID созданного маршрута
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateRouteResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      tags:
        - routes
      summary: Обновить метаинформацию маршрута
      operationId: updateRouteById
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRouteRequest'
      responses:
        '200':
          description: ID обновленного маршрута
          content:
            application/json:
              schema:
                type: string
                format: uuid
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

  /routes/my:
    get:
      tags:
        - routes
      summary: Получить маршруты текущего пользователя
      operationId: getMyRoutes
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 1)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Маршруты текущего пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoutesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /routes/{id}:
    get:
      tags:
        - routes
      summary: Получить полную информацию о маршруте
      operationId: getRouteById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали маршрута
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteFullResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    delete:
      tags:
        - routes
      summary: Удалить свой неопубликованный маршрут
      operationId: deleteRoute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: version
          in: query
          required: false
          description: Ожидаемая версия маршрута для оптимистичной блокировки
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Маршрут удален
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

  /favorites:
    get:
      tags:
        - favorites
      summary: Получить избранные маршруты пользователя с пагинацией
      operationId: getFavorites
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 1)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Список избранного с метаданными пагинации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoutesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  /favorites/{routeId}:
    post:
      tags:
        - favorites
      summary: Добавить маршрут в избранное
      operationId: addFavorite
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Добавлено
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

    delete:
      tags:
        - favorites
      summary: Удалить из избранного
      operationId: removeFavorite
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Удалено
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /routes/{routeId}/reviews:
    get:
      tags:
        - reviews
      summary: Получить список отзывов к маршруту с пагинацией
      operationId: getReviewsByRoute
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          description: Номер страницы (начиная с 1)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Количество отзывов на странице
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Список отзывов с метаданными пагинации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedReviewsResponse'
    post:
      tags:
        - reviews
      summary: Добавить отзыв к маршруту
      operationId: createReview
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
      responses:
        '201':
          description: Отзыв добавлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'

  /routes/{routeId}/reviews/{reviewId}:
    delete:
      tags:
        - reviews
      summary: Удалить отзыв
      operationId: deleteReview
      parameters:
        - name: routeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Удалено
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'


  /admin/users:
    get:
      tags:
        - admin
      summary: Список всех пользователей
      operationId: getAllUsersAdmin
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 1)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  /admin/routes:
    get:
      tags:
        - admin
      summary: Получить маршруты для модерации
      operationId: getAdminRoutes
      parameters:
        - name: page
          in: query
          description: Номер страницы (начиная с 1)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Количество элементов на странице
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/RouteStatus'
      responses:
        '200':
          description: Список маршрутов для модерации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRoutesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    put:
      tags:
        - admin
      summary: Одобрить маршрут
      operationId: approveRoute
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRouteRequest'
      responses:
        '201':
          description: UUID маршрута
          content:
            application/json:
              schema:
                type: string
                format: uuid
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

    delete:
      tags:
        - admin
      summary: Удалить маршрут администратором
      operationId: deleteRouteAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeleteRouteRequest'
      responses:
        '204':
          description: Маршрут удален
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID

  responses:
    BadRequestError:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Ошибка валидации полей формы
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
    UnauthorizedError:
      description: Необходима авторизация (Токен отсутствует или истек)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Недостаточно прав (Нужна роль ADMIN или авторство)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ConflictError:
      description: Конфликт (например, email уже занят, или уже в избранном)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    PaginationMeta:
      type: object
      properties:
        totalItems:
          type: integer
          description: Всего элементов в базе
          example: 154
        totalPages:
          type: integer
          description: Общее количество страниц
          example: 8
        currentPage:
          type: integer
          description: Текущая страница
          example: 1
        itemsPerPage:
          type: integer
          description: Количество элементов на странице
          example: 20

    PaginatedRoutesResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RoutePreviewResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    PaginatedReviewsResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReviewResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Ресурс не найден"
        path:
          type: string
          example: "/api/routes/12"

    ValidationErrorResponse:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Ошибка валидации входных данных"
        path:
          type: string
        validationErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorDto'

    FieldErrorDto:
      type: object
      properties:
        field:
          type: string
          example: "title"
        message:
          type: string
          example: "Название должно содержать минимум 5 символов"

    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: "^[a-zA-Z0-9_]+$"
        email:
          type: string
          format: email
          maxLength: 100
        password:
          type: string
          minLength: 8
          maxLength: 64

    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN

    CheckEmailExistsResponse:
      type: object
      required:
        - exists
      properties:
        exists:
          type: boolean
          example: true

    AuthValidationResponse:
      type: object
      required:
        - userId
        - role
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum:
            - ROLE_USER
            - ROLE_ADMIN

    RouteStatus:
      type: string
      enum:
        - DRAFT
        - PENDING
        - PUBLISHED

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
          format: date-time

    PaginatedUserResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    LocationDto:
      type: object
      required:
        - latitude
        - longitude
        - orderIndex
      properties:
        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
        orderIndex:
          type: integer
          minimum: 0

    CreateRouteRequest:
      type: object
      required:
        - title
        - locations
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 100
        description:
          type: string
          maxLength: 2000
        imageUrl:
          type: string
          format: uri
        locations:
          type: array
          minItems: 2
          maxItems: 50
          items:
            $ref: '#/components/schemas/LocationDto'

    CreateRouteResponse:
      type: object
      required:
        - uuid
        - imageUrl
      properties:
        uuid:
          type: string
          format: uuid
        imageUrl:
          type: string
          format: uri

    UpdateRouteRequest:
      type: object
      required:
        - title
      properties:
        uuid:
          type: string
          format: uuid
        version:
          type: integer
          format: int64
        title:
          type: string
          minLength: 5
          maxLength: 100
        description:
          type: string
          maxLength: 2000

    DeleteRouteRequest:
      type: object
      required:
        - uuid
      properties:
        uuid:
          type: string
          format: uuid
        version:
          type: integer
          format: int64

    RoutePreviewResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        distance:
          type: number
          format: double
          minimum: 0
        durationMinutes:
          type: integer
          minimum: 0
        authorName:
          type: string
        status:
          $ref: '#/components/schemas/RouteStatus'
        is_liked:
          type: boolean

    RouteFullResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        version:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        distance:
          type: number
          format: double
        durationMinutes:
          type: integer
        authorName:
          type: string
        status:
          $ref: '#/components/schemas/RouteStatus'
        locations:
          type: array
          items:
            $ref: '#/components/schemas/LocationDto'
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/ReviewResponse'
        geometry:
          type: string
        is_liked:
          type: boolean

    ReviewRequest:
      type: object
      required:
        - rating
        - comment
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          minLength: 2
          maxLength: 1000

    ReviewResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        authorName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time